Fix pytorch related errors

+++ internal/geopoly.py
@@ -119,6 +119,5 @@ def generate_basis(base_shape,
     # Remove elements of `verts` that are reflections of each other.
     match = compute_sq_dist(verts.T, -verts.T) < eps
     verts = verts[np.any(np.triu(match), 1), :]
-
-  basis = verts[:, ::-1]
+  basis = verts[:, ::-1].astype(np.float32)
   return basis
